SKARA-1696
