+++ bots/common/src/main/java/module-info.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019, 2022, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2022, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -20,21 +20,20 @@
  * or visit www.oracle.com if you need additional information or have any
  * questions.
  */
-import org.openjdk.skara.bots.pr.PullRequestBotFactory;
 
-module org.openjdk.skara.bots.pr {
+/**
+ * The bots.common module is meant for application level logic that needs to be
+ * shared between multiple bots. This is needed for functionality that ties
+ * together multiple different libraries that we don't want to create
+ * dependencies between.
+ */
+module org.openjdk.skara.bots.common {
     requires org.openjdk.skara.vcs;
-    requires org.openjdk.skara.jcheck;
-    requires org.openjdk.skara.host;
-    requires org.openjdk.skara.census;
-    requires org.openjdk.skara.ini;
-    requires org.openjdk.skara.bot;
-    requires org.openjdk.skara.process;
-    requires org.openjdk.skara.email;
-    requires org.openjdk.skara.jbs;
+    requires transitive org.openjdk.skara.forge;
     requires org.openjdk.skara.network;
-    requires org.openjdk.skara.bots.common;
+    requires transitive org.openjdk.skara.jbs;
+    requires transitive org.openjdk.skara.jcheck;
     requires java.logging;
 
-    provides org.openjdk.skara.bot.BotFactory with PullRequestBotFactory;
+    exports org.openjdk.skara.bots.common;
 }
