+++ settings.gradle
@@ -59,6 +59,7 @@ include 'bots:bridgekeeper'
 include 'bots:censussync'
 include 'bots:checkout'
 include 'bots:cli'
+include 'bots:common'
 include 'bots:csr'
 include 'bots:jep'
 include 'bots:forward'
